Fix: adding utils
