+++ index.js
@@ -1,5 +1,6 @@
 const http = require('node:http')
 const fs = require('node:fs')
+const { methodNotAllowed, notFoud } = require('./utils/index')
 
 const desiredPort = process.env.PORT ?? 1234
 
@@ -63,18 +64,6 @@ function startServer () {
       default:
         methodNotAllowed(res)
     }
-  }
-
-  function notFoud (response) {
-    response.setHeader('Content-Type', 'text/html; charset=utf-8')
-    response.statusCode = 404
-    response.end('<h1>404</h1>')
-  }
-
-  function methodNotAllowed (response) {
-    response.setHeader('Content-Type', 'text/html; charset=utf-8')
-    response.statusCode = 405
-    response.end('<h1>405 Method Not Allowed</h1>')
   }
 
   const server = http.createServer(processingRequests)
